// src/types.ts
export type StudentNavTab =
  | 'home'
  | 'materials'
  | 'videos'
  | 'tests'
  | 'profile'
  | 'dashboard'
  | 'results'
  | 'fees';
// ...
export type AdminNavTab =
  | 'overview'
  | 'materials'
  | 'videos'
  | 'tests'
  | 'results'
  | 'fees'
  | 'students'
  | 'announcements';